[css-typed-om] Fix unsupported values in non-list-valued properties.
Currently, when we reify a value, we try to first reify it as a single
value. If that fails, we try reifying it as a list value. This code
was written with the incorrect assumption that only list-valued
properties are stored as CSSValueLists. This is not true, as many single
valued properties (such as grid-auto-flow) are stored as CSSValueLists.

This patch adds an additional check so that only properties that are
explicitly marked as list-valued are reified as list values.

Bug: 829881
Change-Id: I965e38355b9859a43fcb1b940f19237f9f07fdf6